add a basic smart beaconing algorithm
This is a smart beconing algorithm based on the speed and course of a
tracked object. It works by recording the speed and course of the last
broadcast beacon and comparing current stats.

The algorithm calculates an interval between `min_interval` and
`max_interval` seconds, based on object speed and course.

This algorithm was tuned for highway travel in a passenger vehicle.